docs(date-picker): remove tooltip sticker sheet story from filter base buttons
