--- packages/date-picker/docs/filter-button-foundations/filter-base-buttons.stories.tsx
@@ -1,75 +1,0 @@
-import React from "react"
-import { Story } from "@storybook/react"
-import { withDesign } from "storybook-addon-designs"
-import { StoryWrapper } from "../../../../storybook/components/StoryWrapper"
-import {
-  CATEGORIES,
-  SUB_CATEGORIES,
-  SUB_COMPONENTS_FOLDER_NAME,
-} from "../../../../storybook/constants"
-import { FilterBaseButton } from "../../src/FilterDateRangePicker/components/Trigger/FilterBaseButton"
-import { FilterBaseTooltipButton } from "../../src/FilterDateRangePicker/components/Trigger/FilterBaseTooltipButton"
-
-export default {
-  title: `${CATEGORIES.components}/${SUB_CATEGORIES.datePicker}/Filter Date Range Picker/${SUB_COMPONENTS_FOLDER_NAME}/Filter Base Buttons`,
-  subcomponents: {
-    FilterBaseButton,
-    FilterBaseTooltipButton,
-  },
-  parameters: {
-    docs: {
-      description: {
-        component:
-          "This is a subcomponent - use FilterTriggerButton or RemovableFilterTriggerButton.",
-      },
-    },
-  },
-  decorators: [withDesign],
-}
-
-const StickerSheetTemplate: Story = () => (
-  // For Chromatic
-  <div style={{ paddingBottom: "2rem" }}>
-    <StoryWrapper>
-      <StoryWrapper.RowHeader headings={["Base", "Hover", "Active", "Focus"]} />
-      <StoryWrapper.Row rowTitle="Filter Base Button">
-        <div>
-          <FilterBaseButton>Label</FilterBaseButton>
-        </div>
-        <div>
-          <FilterBaseButton classNameOverride="story__filter-button--hover">
-            Label
-          </FilterBaseButton>
-        </div>
-        <div>
-          <FilterBaseButton classNameOverride="story__filter-button--active">
-            Label
-          </FilterBaseButton>
-        </div>
-        <div>
-          <FilterBaseButton classNameOverride="story__filter-button--focus">
-            Label
-          </FilterBaseButton>
-        </div>
-      </StoryWrapper.Row>
-
-      <StoryWrapper.Row rowTitle="Filter Base Tooltip Button">
-        <div>
-          <FilterBaseTooltipButton
-            tooltipText="Tooltip"
-            isTooltipInitiallyVisible
-          >
-            Label
-          </FilterBaseTooltipButton>
-        </div>
-      </StoryWrapper.Row>
-    </StoryWrapper>
-  </div>
-)
-
-export const StickerSheetDefault = StickerSheetTemplate.bind({})
-StickerSheetDefault.storyName = "Sticker Sheet (Default)"
-StickerSheetDefault.parameters = {
-  chromatic: { disable: false },
-  controls: { disable: true },
-}
